Added options to select apps and compare counties
