+++ apps/configs.py
@@ -1,0 +1,1 @@
+STATES_FOLDER = "data/states/"
